+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 0.6 (8/11/2014)
+
++	Added action type for sending a basic implicit intent, example: *INTENT=android.intent.action.VIEW**http://google.com
++	Added action type for executing a Tasker task, example: *TASKER=Play A Playlist**Dubstep Favorites
+
+
 # Version 0.5 (8/11/2014)
 
 +	Fixed logic error preventing connections